Start to refactor work to use wrapped arrays
The initial development branch passed arrays directly into BLAS via the
standard Ada-FORTRAN interface. This worked fine, but has significant
limitations because many of the most interesting algorithms need or at
least benefit from being able to do calculations on parts of vectors or
matrices - this is why the BLAS API takes separate dimensions and stride
parameters for matrices.

In this branch I will look at an alternative approach where the
underlying data is hidden inside a record which has primitive operations
to return suitable stride and size parameters, and to return a pointer
to the aliased data itself. This will allow "view" or "slice" types to
be created that point to pre-existing data in memory, but with different
stride or length parameters. This assumes that the FORTRAN ABI for
passing arrays is the same as the C ABI, as we need to use the pointer
arithmetic in Interfaces.C.Pointers.