Removed use of Interfaces.C.Pointers
Interfaces.C.Pointers is not actually necessary so can be removed. It is
now assumed that the FORTRAN ABI for array accesses is by passing a
pointer to the first item in the array - this is likely to be true on
virtually every platform, but is not guaranteed AFAIK.

There are now two type of pointers for each of Vectors and Matrices, so
pointers used as "in" parameters for FORTRAN procedures can be
differentiated from those used for "out" parameters.